requirements.txt
backcall==0.2.0
black==21.12b0
cfgv==3.3.1
click==8.0.3
cycler==0.11.0
decorator==5.1.0
distlib==0.3.4
filelock==3.4.2
flake8==4.0.1
fonttools==4.28.4
identify==2.4.1
ipdb==0.13.9
ipython==7.30.1
jedi==0.18.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.21.4
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
pre-commit==2.16.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scipy==1.7.3
six==1.16.0
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
typing-extensions==4.0.1
virtualenv==20.11.2
wcwidth==0.2.5
